Fix DIVIDE (not an integer divide).

+++ core/model/src/main/java/it/unibz/inf/ontop/model/type/impl/NonIntegerNumericInferenceRule.java
@@ -5,6 +5,7 @@ import it.unibz.inf.ontop.model.type.TermType;
 import java.util.Optional;
 
 import static it.unibz.inf.ontop.model.OntopModelSingletons.TYPE_FACTORY;
+import static it.unibz.inf.ontop.model.type.COL_TYPE.DECIMAL;
 import static it.unibz.inf.ontop.model.type.COL_TYPE.INTEGER;
 import static it.unibz.inf.ontop.model.type.COL_TYPE.INTEGER_TYPES;
 
@@ -17,6 +18,6 @@ public class NonIntegerNumericInferenceRule extends NumericTermTypeInferenceRule
     protected Optional<TermType> postprocessInferredType(Optional<TermType> optionalTermType) {
         // No need to call super.postprocessInferredType()
         return optionalTermType
-                .map(t -> INTEGER_TYPES.contains(t.getColType()) ? TYPE_FACTORY.getTermType(INTEGER) : t);
+                .map(t -> INTEGER_TYPES.contains(t.getColType()) ? TYPE_FACTORY.getTermType(DECIMAL) : t);
     }
 }
